[BROKEN] Compressed mcdata, better build, etc

+++ scripts/build.sh
@@ -1,28 +1,39 @@
 #!/bin/sh
 #CWD: Project root
 
+if [ $(basename "$PWD") = "scripts" ]; then
+  echo "Error: Process CWD must be project root"
+  exit 1
+fi
+
 rm -r -f ./dist
 rm -r -f ./build
+mkdir ./build
+mkdir ./build/game-versions
 npm i --no-fund
 npx tsc
-powershell -File ./scripts/generate-mcdata-index-patch.ps1
-powershell -File ./scripts/generate-mcversions-table.ps1
-powershell -File ./scripts/generate-version-dependencies-table.ps1
+node ./scripts/generate-compressed-mcdata.js
+node ./scripts/generate-mcdata-index-patch.js
+node ./scripts/generate-mcversions-table.js
 
 #Little patch
-cp ./node_modules/minecraft-data/data.js ./~__minecraft-data--data.js
-cp ./src/launcher/_mcdata-index.cjs ./node_modules/minecraft-data/data.js
+cp ./node_modules/minecraft-data/data.js ./build/~original--mcdata-index.js #Save origin
+cp ./src/launcher/_mcdata-provider.cjs ./node_modules/minecraft-data/data.js #Inject modified
+cp ./dist/mcdata-real-index.cjs ./node_modules/minecraft-data/_real-index.js
 
 #Bundle launcher
-mkdir ./build
 cp -r ./src/launcher/program ./build/program
+cp -r ./dist/compressed-mcdata/pc ./build/game-versions/pc
 cp ./src/launcher/ReleaseReadme.md ./build/README.md
 cp ./LICENSE ./build/LICENSE
 cp ./package.json ./build/program/package.json
+cp ./dist/mcversions-index.json ./build/program/
 
 #Bundle core & dependencies
-npx esbuild ./dist/core/init_bot.js --bundle --outfile=./build/program/mca-bundle.cjs --format=cjs --platform=node
+npx esbuild ./dist/core/init_bot.js --bundle \
+  --outfile=./build/program/mca-bundle.cjs --format=cjs --platform=node
 
 #Revert patch
-cp ./~__minecraft-data--data.js ./node_modules/minecraft-data/data.js
-rm ./~__minecraft-data--data.js
+cp ./build/~original--mcdata-index.js ./node_modules/minecraft-data/data.js
+rm ./build/~original--mcdata-index.js
+rm ./node_modules/minecraft-data/_real-index.js
